src/agents/a2a_hello_world/agent.py
"""This module defines a simple agent that can get the weather and time."""
import datetime
from zoneinfo import ZoneInfo
from google.adk.agents import Agent
from google.adk.a2a.utils.agent_to_a2a import to_a2a


def get_weather(city: str) -> dict:
    """Retrieves the current weather report for a specified city.

    Args:
        city (str): The name of the city for which to retrieve the weather report.

    Returns:
        dict: status and result or error msg.
    """
    if city.lower() == "new york":
        return {
            "status": "success",
            "report": (
                "The weather in New York is sunny with a temperature of 25 degrees"
                " Celsius (77 degrees Fahrenheit)."
            ),
        }
    return {
        "status": "error",
        "error_message": f"Weather information for '{city}' is not available.",
    }


def get_current_time(city: str) -> dict:
    """Returns the current time in a specified city.

    Args:
        city (str): The name of the city for which to retrieve the current time.

    Returns:
        dict: status and result or error msg.
    """

    if city.lower() != "new york":
        return {
            "status": "error",
            "error_message": (
                f"Sorry, I don't have timezone information for {city}."
            ),
        }
    tz_identifier = "America/New_York"
    tz = ZoneInfo(tz_identifier)
    now = datetime.datetime.now(tz)
    report = (
        f'The current time in {city} is {now.strftime("%Y-%m-%d %H:%M:%S %Z%z")}'
    )
    return {"status": "success", "report": report}


root_agent = Agent(
    name="weather_time_agent",
    model="gemini-2.5-flash",
    description=(
        "Agent to answer questions about the time and weather in a city."
    ),
    instruction=(
        "You are a helpful agent who can answer user questions about the time "
        "and weather in a city."
    ),
    tools=[get_weather, get_current_time],
)

if __name__ == "__main__":
    import uvicorn
    a2a_app = to_a2a(root_agent, port=8080)
    # Use host='0.0.0.0' to allow external access.
    uvicorn.run(a2a_app, host='0.0.0.0', port=8080)
