requirements.txt
# Dev environment
pip
autopep8

# App
google-adk
google-cloud-aiplatform
